Fix warnings.

+++ Dev/Cpp/Effekseer/Effekseer/Culling/Culling3D.ObjectInternal.h
@@ -38,7 +38,7 @@ namespace Culling3D
 				radius = 0.0f;
 				if (Type == OBJECT_SHAPE_TYPE_NONE) radius = 0.0f;
 				if (Type == OBJECT_SHAPE_TYPE_SPHERE) radius = Data.Sphere.Radius;
-				if (Type == OBJECT_SHAPE_TYPE_CUBOID) radius = sqrt(Data.Cuboid.X * Data.Cuboid.X + Data.Cuboid.Y * Data.Cuboid.Y + Data.Cuboid.Z * Data.Cuboid.Z) / 2.0f;
+				if (Type == OBJECT_SHAPE_TYPE_CUBOID) radius = sqrtf(Data.Cuboid.X * Data.Cuboid.X + Data.Cuboid.Y * Data.Cuboid.Y + Data.Cuboid.Z * Data.Cuboid.Z) / 2.0f;
 			}
 
 			float GetRadius()
