+++ Dev/Cpp/Effekseer/Effekseer/Effekseer.EffectNodeTrack.cpp
@@ -151,8 +151,8 @@ void EffectNodeTrack::Rendering(const Instance& instance, const Instance* next_i
 	if (renderer != NULL)
 	{
 		float t = (float)instance.m_LivingTime / (float)instance.m_LivedTime;
-		int32_t time = instance.m_LivingTime;
-		int32_t livedTime = instance.m_LivedTime;
+		int32_t time = (int32_t)instance.m_LivingTime;
+		int32_t livedTime = (int32_t)instance.m_LivedTime;
 
 		SetValues(m_instanceParameter.ColorLeft, instance, m_currentGroupValues.ColorLeft, TrackColorLeft, time, livedTime);
 		SetValues(m_instanceParameter.ColorCenter, instance, m_currentGroupValues.ColorCenter, TrackColorCenter, time, livedTime);
@@ -212,8 +212,8 @@ void EffectNodeTrack::InitializeRenderedInstanceGroup(InstanceGroup& instanceGro
 void EffectNodeTrack::InitializeRenderedInstance(Instance& instance, Manager* manager)
 {
 	// Calculate only center
-	int32_t time = instance.m_LivingTime;
-	int32_t livedTime = instance.m_LivedTime;
+	int32_t time = (int32_t)instance.m_LivingTime;
+	int32_t livedTime = (int32_t)instance.m_LivedTime;
 
 	Color c;
 	SetValues(c, instance, m_currentGroupValues.ColorCenterMiddle, TrackColorCenterMiddle, time, livedTime);
@@ -232,8 +232,8 @@ void EffectNodeTrack::InitializeRenderedInstance(Instance& instance, Manager* ma
 void EffectNodeTrack::UpdateRenderedInstance(Instance& instance, Manager* manager)
 {
 	// Calculate only center
-	int32_t time = instance.m_LivingTime;
-	int32_t livedTime = instance.m_LivedTime;
+	int32_t time = (int32_t)instance.m_LivingTime;
+	int32_t livedTime = (int32_t)instance.m_LivedTime;
 
 	Color c;
 	SetValues(c, instance, m_currentGroupValues.ColorCenterMiddle, TrackColorCenterMiddle, time, livedTime);
